+++ capturecat-core/src/main/java/com/capturecat/core/CoreExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.capturecat;
+package com.capturecat.core;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
