+++ capturecat-core/src/main/java/com/capturecat/core/HealthCheckController.java
@@ -1,4 +1,4 @@
-package com.capturecat;
+package com.capturecat.core;
 
 import java.util.HashMap;
 import java.util.Map;
