+++ capturecat-core/src/main/java/com/capturecat/core/config/SecurityConfig.java
@@ -1,4 +1,4 @@
-package com.capturecat.config;
+package com.capturecat.core.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
