+++ capturecat-core/src/test/java/com/capturecat/core/HealthCheckControllerTest.java
@@ -1,4 +1,4 @@
-package com.capturecat;
+package com.capturecat.core;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
